Fix up a lot from the messed merge
