+++ WorldExplorer/WorldExplorer/Extensions/StringExtensions.cs
@@ -1,0 +1,10 @@
+﻿namespace WorldExplorer
+{
+    public static class StringExtensions
+    {
+        public static string TrimQuotes(this string value)
+        {
+            return value.Replace("\"", "");
+        }
+    }
+}
